LPS-109189 Create BundleUtil class to make IndexUpdaterUtil API more cohesive. Now is performing OSGi stuff and indexes related stuff.
