+++ portal-impl/src/com/liferay/portal/util/BundleUtil.java
@@ -1,0 +1,85 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.util;
+
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.StringUtil;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+import java.net.URL;
+
+import java.util.Dictionary;
+
+import org.osgi.framework.Bundle;
+import org.osgi.framework.BundleContext;
+
+/**
+ * @author Mariano Álvaro Sáiz
+ */
+public class BundleUtil {
+
+	public static Bundle getBundle(
+		BundleContext bundleContext, String bundleSymbolicName) {
+
+		for (Bundle bundle : bundleContext.getBundles()) {
+			if (bundleSymbolicName.equals(bundle.getSymbolicName())) {
+				return bundle;
+			}
+		}
+
+		throw new IllegalArgumentException(
+			"Module with symbolic name " + bundleSymbolicName +
+				" does not exist");
+	}
+
+	public static String getSQLTemplateString(
+		Bundle bundle, String templateName) {
+
+		URL resource = bundle.getResource("/META-INF/sql/" + templateName);
+
+		if (resource == null) {
+			if (_log.isDebugEnabled()) {
+				_log.debug("Unable to locate SQL template " + templateName);
+			}
+
+			return null;
+		}
+
+		try (InputStream inputStream = resource.openStream()) {
+			return StringUtil.read(inputStream);
+		}
+		catch (IOException ioException) {
+			_log.error(
+				"Unable to read SQL template " + templateName, ioException);
+
+			return null;
+		}
+	}
+
+	public static boolean isLiferayServiceBundle(Bundle bundle) {
+		Dictionary<String, String> headers = bundle.getHeaders(
+			StringPool.BLANK);
+
+		return GetterUtil.getBoolean(headers.get("Liferay-Service"));
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(BundleUtil.class);
+
+}
